Check and fix logging formats
Ignore-this: 3cc986f7ec895d08000d1ca44630c6a4

darcs-hash:e2c77541803c57afc808510576fa7e8d4186d9ee

+++ source/clients/qmon/qmon_cq.c
@@ -1655,7 +1655,7 @@ lListElem *qep
 
    sge_strlcat(info, "\n", sizeof(info)); 
 
-   DPRINTF(("info is %d long\n", strlen(info)));
+   DPRINTF(("info is %d long\n", (int) strlen(info)));
    
 #undef WIDTH
    DEXIT;
